paas/application-manager/src/main/java/eu/ascetic/paas/applicationmanager/amonitor/model/parser/Parser.java
package eu.ascetic.paas.applicationmanager.amonitor.model.parser;

import java.io.IOException;

import org.codehaus.jackson.JsonGenerationException;
import org.codehaus.jackson.map.JsonMappingException;
import org.codehaus.jackson.map.ObjectMapper;
import org.codehaus.jackson.map.SerializationConfig;
import org.codehaus.jackson.map.annotate.JsonSerialize.Inclusion;

import eu.ascetic.paas.applicationmanager.amonitor.model.EnergyCosumed;

/**
 * 
 * Copyright 2014 ATOS SPAIN S.A. 
 * 
 * Licensed under the Apache License, Version 2.0 (the License);
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 * 
 * http://www.apache.org/licenses/LICENSE-2.0
 * 
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 * 
 * @author: David Garcia Perez. Atos Research and Innovation, Atos SPAIN SA
 * e-mail: david.garciaperez@atos.net 
 * 
 * Parser to generate the strings for the application monitor
 */
public class Parser {

	public static String getJSONEnergyConsumed(EnergyCosumed energyCosumed) throws JsonGenerationException, JsonMappingException, IOException {

		ObjectMapper mapper = new ObjectMapper();
		mapper.setSerializationInclusion(Inclusion.NON_EMPTY);
        mapper.enable(SerializationConfig.Feature.INDENT_OUTPUT);
		
        return mapper.writeValueAsString(energyCosumed);
	}
}
